chore(mcp-app): remove the session prune and reclamation ops tooling
The prune backfill and the reclamation A/B are both done, so drop the
driver script, /admin/prune, the /admin/inspect, /admin/wipe and
/admin/config diagnostics, and the MCP_PRUNE_ADMIN_TOKEN secret gating
them.

The 7-day self-expiry and the destroy() abort override stay: prune.ts
folds into worker.ts as expiry-local code, minus the dry-run and force
branches only the endpoint used.
