+++ apps/mcp-app/package.json
@@ -14,8 +14,6 @@
 		"dev:tunnel": "yarn build && bash dev-tunnel.sh",
 		"deploy": "yarn build && wrangler deploy",
 		"lint": "yarn run -T tsx ../../internal/scripts/lint.ts",
-		"prune:list": "tsx scripts/prune-sessions.ts list",
-		"prune:run": "tsx scripts/prune-sessions.ts prune",
 		"test": "yarn run -T vitest --passWithNoTests",
 		"test-ci": "yarn run -T vitest run --passWithNoTests",
 		"test-coverage": "yarn run -T vitest run --coverage --passWithNoTests"
